tests/requirements-test.txt
appdirs==1.4.4
astroid==3.0.2
aws-lambda-context==1.1.0
beautifulsoup4==4.12.3
black==23.12.1
boto3==1.34.23
boto3-stubs==1.34.25
botocore==1.34.23
botocore-stubs==1.34.25
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.7
distlib==0.3.8
filelock==3.13.1
frozendict==2.4.0
html5lib==1.1
identify==2.5.33
idna==3.6
isort==5.13.2
jmespath==1.0.1
joblib==1.3.2
lxml==5.1.0
mccabe==0.7.0
multitasking==0.0.11
mypy==1.8.0
mypy-boto3-cloudformation==1.34.0
mypy-boto3-dynamodb==1.34.23
mypy-boto3-ec2==1.34.24
mypy-boto3-lambda==1.34.0
mypy-boto3-rds==1.34.24
mypy-boto3-s3==1.34.14
mypy-boto3-sqs==1.34.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pandas-stubs==2.1.4.231227
pathspec==0.12.1
peewee==3.17.0
platformdirs==4.1.0
pre-commit==3.6.0
pyarrow==15.0.0
pylint==3.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
scikit-learn==1.3.2
scipy==1.12.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
threadpoolctl==3.2.0
tomlkit==0.12.3
tqdm==4.66.1
types-awscrt==0.20.2
types-pytz==2023.3.1.1
types-requests==2.31.0.20240125
types-s3transfer==0.10.0
types-tqdm==4.66.0.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
virtualenv==20.25.0
webencodings==0.5.1
yfinance==0.2.36
